tests/requirements.txt
Pillow
pandas
scipy
#opencv-python
#pylibjpeg
#torchjpeg